requirements.txt
matplotlib==3.4.2
pandas==1.3.5
python-dotenv==0.20.0
requests==2.25.1
